3.5.0 dev

+++ Examples/core/core_input_mouse_wheel.cs
@@ -37,7 +37,7 @@ namespace Examples
             {
                 // Update
                 //----------------------------------------------------------------------------------
-                boxPositionY -= (GetMouseWheelMove() * scrollSpeed);
+                boxPositionY -= (int)(GetMouseWheelMove() * scrollSpeed);
                 //----------------------------------------------------------------------------------
 
                 // Draw
